% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esomnibus.R
\name{esomnibus.default}
\alias{esomnibus.default}
\title{Calculate Epps-Singleton two-sample test}
\usage{
\method{esomnibus}{default}(
  y1 = numeric(),
  y2,
  t = c(0.4, 0.8),
  iqr_type = 7,
  small_correction = FALSE
)
}
\arguments{
\item{y1}{Vector of data from first sample}

\item{y2}{Vector of data from second sample}

\item{t}{Vector of points to calculate empirical characteristic function at. Default c(0.4,0.8)}

\item{iqr_type}{How to calculate semi-inter-quartile-range. Reference the type argument in the IQR function. Defaults to 7. Special value: 99, as referenced in the appendix to the original publication.}

\item{small_correction}{Should the small sample correction be applied? (TRUE or FALSE)}
}
\value{
A list (which is also of class 'esomnibustest') with the following elements:
}
\description{
Calculate Epps-Singleton two-sample test
}
