% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esomnibus.R
\name{esomnibus}
\alias{esomnibus}
\alias{esomnibus.default}
\alias{esomnibus.formula}
\title{Epps-Singleton two-sample test for equality of distribution}
\usage{
esomnibus(x, ...)
}
\arguments{
\item{x}{A vector or a formula}

\item{...}{}
}
\value{
An S3 object of type
}
\description{
Epps-Singleton two-sample test for equality of distribution
}
\examples{
x = runif(50)
y = rnorm(50)
esomnibus(x,y)
}
