% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esomnibus.R
\name{esomnibus.test.default}
\alias{esomnibus.test.default}
\title{Calculate Epps-Singleton two-sample test}
\usage{
\method{esomnibus.test}{default}(y1, y2, t = c(0.4, 0.8), iqr_type = 7, small_correction = FALSE)
}
\arguments{
\item{y1}{Vector of data from first sample}

\item{y2}{Vector of data from second sample}

\item{t}{Vector of points to calculate empirical characteristic function at. Default c(0.4,0.8)}

\item{iqr_type}{How to calculate semi-inter-quartile-range. Reference the type argument in the IQR function. Defaults to 7. Special value: 99, as referenced in the appendix to the original publication.}

\item{small_correction}{Should the small sample correction be applied? (TRUE or FALSE)}
}
\value{
An S3 object of type estest
}
\description{
Calculate Epps-Singleton two-sample test
}
\examples{
x = runif(50)
y = rnorm(50)
esomnibus.test.default(x,y)
}
